requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-dotenv
openai
httpx
APScheduler
pytest
